♻️ Input 컴포넌트를 NormalInput 컴포넌트로 이동시키고, variant로 분기
