+++ src/components/Input/Input.types.ts
@@ -1,4 +1,16 @@
-interface InputBaseType {}
+import { type InputHTMLAttributes } from 'react';
+
+interface InputBaseType {
+  variant?: 'normal' | 'drop-down';
+}
+
+export interface NormalInputType extends InputBaseType, Omit<InputHTMLAttributes<HTMLInputElement>, 'onChange'> {
+  variant?: 'normal';
+  value: string;
+  onChange?: (value: string) => void;
+
+  description: string;
+}
 
 export interface DropdownValueType {
   imgUrl?: string;
@@ -7,11 +19,13 @@ export interface DropdownValueType {
 }
 
 export interface DropDownInputType extends InputBaseType {
+  variant: 'drop-down';
   selected: DropdownValueType | null;
   onSelect: (value: DropdownValueType) => void;
   list: DropdownValueType[];
   placeholder?: string;
-
   title: string;
   required?: boolean;
 }
+
+export type InputType = DropDownInputType | NormalInputType;
